chore(Cleanup) General ui cleanup before go live
