+++ apps/web/app/[locale]/featureFlags/layout.tsx
@@ -1,6 +1,14 @@
+import { PgSessions } from "auth/sessions";
 import FeatureFlagsMenu from "./FeatureFlagsMenu";
+import { RedirectType, redirect } from "next/navigation";
 
-export default ({ children }) => {
+export default async ({ children }) => {
+  const { publicServant } = await PgSessions.get();
+
+  if (!publicServant) {
+    return redirect("/", RedirectType.replace);
+  }
+
   return (
     <div
       style={{
